requirements.txt
streamlit>=1.30
pandas>=2.1
numpy>=1.26
scipy>=1.11
scikit-learn>=1.3
plotly>=5.18
matplotlib>=3.8
seaborn>=0.13
pyyaml>=6.0
requests>=2.31
joblib>=1.3
python-dateutil>=2.8

# Critical for Tabular ML Page
xgboost>=2.0
lightgbm>=4.3

# Optional for AI Insights (Warning: High Memory Usage)
# transformers>=4.40
# torch>=2.2

# Optional for deep learning time-series (TFT)
# pytorch-forecasting>=1.0
# pytorch-lightning>=2.2
